refactor to use Arch instead of Translator trait
