--- symex/src/general_assembly/translators/mod.rs
@@ -1,2 +1,0 @@
-// All the translators for different instruction sets and encodings.
-pub mod armv6_m;
